Inicio de asignacion personal proyecto

+++ src/main/java/com/marioborrego/api/calculodeduccionesbackend/economico/business/interfaces/EconomicoService.java
@@ -10,7 +10,7 @@ import org.springframework.data.domain.Pageable;
 public interface EconomicoService {
     boolean eliminarEconomico(EconomicoListadoGeneralDto economico);
     Page<EconomicoListadoGeneralDto> obtenerEconomicosPaginados(Pageable pageable);
-    boolean comprobarExistenciaEconomico(@NotBlank(message = "El CIF es obligatorio") @Size(min = 9, max = 9, message = "El CIF debe tener 9 caracteres") String cif, @NotNull(message = "La anualidad es obligatoria") Integer anualidad);
+    boolean comprobarExistenciaEconomico(@NotBlank(message = "El CIF es obligatorio") @Size(min = 9, max = 9, message = "El CIF debe tener 9 caracteres") String cif, @NotNull(message = "La anualidad es obligatoria") Long anualidad);
     EconomicoCreadoDTO crearEconomico(CrearEconomicoDTO crearEconomicoDTO);
     EconomicoDTO obtenerEconomico(Long idEconomico);
     void actualizarDatosEconomico(ActualizarDatosEconomicoDTO economico);
